requirements.txt
annotated-types==0.7.0
antiorm==1.2.1
anyio==3.7.1
api==0.0.7
async-generator==1.10
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
db==0.1.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
fastapi-login==1.10.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jose==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
motor==3.4.0
nose==1.3.7
orjson==3.10.3
outcome==1.3.0.post0
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-wordpress-xmlrpc==2.3
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
router==0.1
rsa==4.9
shellingham==1.5.4
sib-api-v3-sdk==7.6.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.37.2
trio==0.21.0
typer==0.12.3
typing_extensions==4.12.0
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.0
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
